rename config subpkg to _config for the time being
